Upload post datastructure modified
